EY-2558: Opnar for å sende ei spesifikk sak inn til migreringa
