+++ apps/etterlatte-migrering/src/main/kotlin/migrering/pen/PenKlient.kt
@@ -4,6 +4,7 @@ import io.ktor.client.HttpClient
 import io.ktor.client.call.body
 import io.ktor.client.request.get
 import io.ktor.http.isSuccess
+import no.nav.etterlatte.migrering.Pesyssak
 import org.slf4j.LoggerFactory
 
 class PenKlient(
@@ -13,15 +14,13 @@ class PenKlient(
     private val logger = LoggerFactory.getLogger(this::class.java)
     suspend fun hentSak(
         sakid: Long
-    ): PenDTO {
+    ): Pesyssak {
         logger.info("Henter sak $sakid fra PEN")
         val sak = pen.get("$resourceUrl/barnepensjon-migrering/grunnlag/$sakid")
         return if (sak.status.isSuccess()) {
-            sak.body<PenDTO>()
+            sak.body<Pesyssak>()
         } else {
             throw RuntimeException("Kunne ikke hente sak $sakid fra PEN")
         }
     }
 }
-
-data class PenDTO(val sakid: Long)
