+++ libs/rapidsandrivers-extras/src/main/kotlin/rapidsandrivers/migrering/Migreringshendelser.kt
@@ -5,6 +5,8 @@ object Migreringshendelser {
     private const val PREFIX = "MIGRERING:"
 
     const val START_MIGRERING = "${PREFIX}START_MIGRERING"
+    const val MIGRER_SPESIFIKK_SAK = "${PREFIX}MIGRER_SPESIFIKK_SAK"
+
     const val MIGRER_SAK = "${PREFIX}MIGRER_SAK"
     const val LAGRE_GRUNNLAG = "${PREFIX}LAGRE_GRUNNLAG"
     const val VILKAARSVURDER = "${PREFIX}VILKAARSVURDER"
